Implemented classes

+++ DataHandling.py
@@ -1,9 +1,11 @@
 import json
+from pathlib import Path
 class DataHandler:
     def __init__(self,file_name):
         self.file_name=file_name
-        with open(self.file_name, "w") as f:
-            json.dump({},f)
+        if not Path(self.file_name).exists():
+            with open(self.file_name, "w") as f:
+                json.dump({},f)
         self.full_map={}
         self.full_df={}
         self.full_idf={}
@@ -18,6 +20,7 @@ class DataHandler:
             del file["phone"]
         return file
     def add_resume(self,key,value):
+
         with open(self.file_name) as f:
             data=json.load(f)
         data[key]=self.clean_file(value)
